+++ backend/src/helpers/VeridaHelper.ts
@@ -1,21 +1,32 @@
 import NodeCache from 'node-cache';
 import { Wallet } from 'ethers';
-import { Context, EnvironmentType, Network } from '@verida/client-ts';
+import { Context, Network } from '@verida/client-ts';
 import { AutoAccount } from '@verida/account-node';
 import { Credentials } from '@verida/verifiable-credentials';
+import EncryptionUtils from '@verida/encryption-utils'
 import { IssuerDto } from '../modules/issuer/dto';
 import { Issuer } from '../modules/issuer/interfaces/issuer.interface';
 import { IssueCredentialDto } from '../modules/credential/dto';
 import { SendMessageResponse } from 'src/models/User';
-import { BLOCK_CHAIN, DURATION_TTL } from '../constants/constant.config';
+import { BLOCK_CHAIN, DURATION_TTL } from '../constants';
+import { config } from 'src/config';
 
-const { VERIDA_APP_NAME, VERIDA_TESTNET_DEFAULT_SERVER } = process.env;
+const appConfig = config()
 
 const CacheManager = new NodeCache();
 
+const VERIDA_DEFAULT_DID_SERVERS = appConfig.veridaDefaultDidServers;
+const VERIDA_DEFAULT_STORAGE_NODES = appConfig.veridaDefaultStorageNodes;
+
+const VERIDA_CREDENTIAL_SCHEMA = 'https://common.schemas.verida.io/credential/base/v0.2.0/schema.json'
+
 export default class VeridaHelper {
-  /**
-   */
+/**
+ * method to create a new issuer account 
+ * @param name 
+ * @param avatarUri 
+ * @returns 
+ */
   static async createAccount(
     name: string,
     avatarUri?: string,
@@ -27,11 +38,13 @@ export default class VeridaHelper {
     address: string;
   }> {
     const account = await VeridaHelper.generateAccount();
+
     const context = await VeridaHelper.connect(account.privateKey);
+
     const did = await context.getAccount().did();
 
     this.setProfile(context, name, avatarUri);
-
+    
     return {
       chain: BLOCK_CHAIN,
       did: did,
@@ -77,57 +90,109 @@ export default class VeridaHelper {
     issuer: Issuer,
     credentialItem: IssueCredentialDto,
   ): Promise<any> {
+    const credentials = new Credentials();
     try {
       const context = await VeridaHelper.connect(issuer.privateKey);
 
-      const credentials = new Credentials();
+      let payload: any = {}
+      const proofStrings: any = {}
+      if (credentialItem.proofs) {
+        payload.proofs = {}
+        
+        for (let key in credentialItem.proofs) {
+          const proofItems = credentialItem.proofs[key]
+          for (let i in proofItems) {
+            const proofItem = proofItems[i]
+            if (proofItem.startsWith('$')) {
+              proofItems[i] = credentialItem.data[proofItem.substring(1)]
+            }
+          }
 
-      const credentialData = await credentials.createCredentialJWT({
+          const proofString = proofItems.join('-')
+          const sig = EncryptionUtils.signData(proofString, new Uint8Array(Buffer.from(issuer.privateKey.slice(2), 'hex')))
+          payload.proofs[key] = sig
+          proofStrings[key] = proofString
+        }
+      }
+
+      const didJwtVc = await credentials.createCredentialJWT({
         context: context as any,
         data: credentialItem.data,
         subjectId: credentialItem.did,
+        schema: credentialItem.schema,
+        payload
       });
 
-      const data = await VeridaHelper.sendMessage(
-        credentialData,
+      // @ts-ignore
+      const generatedCredential = await credentials.verifyCredential(didJwtVc, {})
+
+      // Credential record that could be saved into a database (or sent to a user)
+      const credentialRecord = {
+        schema: VERIDA_CREDENTIAL_SCHEMA,
+        name: credentialItem.name,
+        summary: credentialItem.summary,
+        didJwtVc
+      }
+
+      const messageData = await VeridaHelper.sendMessage(
+        credentialItem.title,
+        credentialRecord,
         context,
         credentialItem.did,
       );
 
-      return data;
+      await context.close()
+
+      return {
+        generatedCredential,
+        messageId: messageData.id,
+        proofStrings
+      };
     } catch (error) {
+      if (credentials.getErrors()) {
+        console.log(credentials.getErrors())
+        return {
+          error: credentials.getErrors()
+        }
+      }
       console.log({ error });
+      return {
+        error: error.message
+      }
     }
   }
 
   static async sendMessage(
-    data: any,
+    subject: string,
+    credentialRecord: any,
     context: any,
     did: string,
   ): Promise<SendMessageResponse> {
     const type = 'inbox/type/dataSend';
-
+    try {
     const config = {
       recipientContextName: 'Verida: Vault',
       did,
     };
 
-    const title = data['title'];
     const messaging = await context.getMessaging();
 
     const messageData = {
-      data: [data],
+      data: [credentialRecord],
     };
 
     const response = await messaging.send(
       did,
       type,
       messageData,
-      title,
+      subject,
       config,
     );
 
     return response as SendMessageResponse;
+    } catch (error) {
+      console.log({error});
+    }
   }
 
   private static async connect(issuerPrivateKey: string): Promise<Context> {
@@ -146,29 +211,38 @@ export default class VeridaHelper {
    */
 
   private static async init(issuerPrivateKey: string) {
+
+    const defaultEndpoints = {
+      defaultDatabaseServer: {
+        type: 'VeridaDatabase',
+        endpointUri: VERIDA_DEFAULT_STORAGE_NODES,
+      },
+      defaultMessageServer: {
+        type: 'VeridaMessage',
+        endpointUri: VERIDA_DEFAULT_STORAGE_NODES,
+      },
+    };
+
+    const account = new AutoAccount(defaultEndpoints, {
+      privateKey: issuerPrivateKey,
+      environment: appConfig.veridaEnvironment,
+      didClientConfig: {
+        callType: 'web3',
+        web3Config: {
+          privateKey: appConfig.polygonPrivateKey,
+          rpcUrl: appConfig.rpcUrl
+        },
+        didEndpoints: VERIDA_DEFAULT_DID_SERVERS,
+      }
+    })
+    
     const context = await Network.connect({
       client: {
-        environment: EnvironmentType.TESTNET,
+        environment: appConfig.veridaEnvironment
       },
-      //@ts-ignore
-      account: new AutoAccount(
-        {
-          defaultDatabaseServer: {
-            type: 'VeridaDatabase',
-            endpointUri: VERIDA_TESTNET_DEFAULT_SERVER,
-          },
-          defaultMessageServer: {
-            type: 'VeridaMessage',
-            endpointUri: VERIDA_TESTNET_DEFAULT_SERVER,
-          },
-        },
-        {
-          privateKey: issuerPrivateKey,
-          environment: EnvironmentType.TESTNET,
-        },
-      ),
+      account,
       context: {
-        name: VERIDA_APP_NAME,
+        name: appConfig.veridaAppName,
       },
     });