+++ backend/src/modules/credential/interfaces/credential.interface.ts
@@ -7,4 +7,5 @@ export interface Credential extends Document {
   credentialId: string;
   revoked: boolean;
   data: object;
+  proofStrings: object;
 }