+++ backend/src/seeder/seeder.module.ts
@@ -6,11 +6,12 @@ import { AdminService } from '../modules/admin/admin.service';
 import { SeederService } from './seeder.service';
 import { AdminModule } from '../modules/admin/admin.module';
 import { AdminSchema } from '../schemas/admin.schema';
+import { config } from 'src/config';
 
 @Module({
     imports: [
         ConfigModule.forRoot(),
-        MongooseModule.forRoot(process.env.DB_URL),
+        MongooseModule.forRoot(config().dbURL),
         MongooseModule.forFeature([
             { name: 'Admin', schema: AdminSchema }
         ]),