+++ CHANGELOG.md
@@ -1,4 +1,12 @@
-# Changelog
 
-# 7-Feb-2021
-* First open source release
+----------------
+
+- Upgrade to Acacia v2.0 of the protocol
+- Credential verification errors are sent in JSON error response
+- Support a "proofs" property that will sign a proof and include it in the data
+@todo: delete context data, so nothing local is retained
+
+2021-02-07
+----------------
+
+- First open source release